+++ README.md
@@ -1,1 +1,5 @@
 # water-history
+
+
+###Useful React References for
+* [Modular App components](https://toddmotto.com/react-create-class-versus-component/)
