+++ src/Components/cover.jsx
@@ -1,0 +1,22 @@
+var React = require('react');
+
+module.exports = function() {
+  return (
+  <div className="container-fluid charcoal-bg">
+    <div className="row cover center">
+      <div className="col-lg-2 col-lg-offset-10 col-md-3 col-md-offset-9 col-sm-3 col-sm-offset-9 col-xs-4 col-xs-offset-8 center paper-bg">
+        <a href="resources.html"><h4>RESOURCES</h4></a>
+      </div>
+      <h1 className="title">West Virginia's Water Crisis</h1>
+      <h2 className="title">The Story of WV American Water</h2>
+    </div>
+    <div className="row">
+      <div className="col-sm-10 col-sm-offset-1 col-lg-8 col-lg-offset-2 top-space bottom-space">
+        <h3>A chemical spill poisoned the drinking water of approximately 300,000 West Virginians on January 9th, 2014...</h3>
+        <p>Many were without potable water for weeks – for some it was months. On January 9th, and still today, drinking water in the 9 county area was coming not from a municipal source, as with most of the United States, but a private corporation – West Virginia American Water. Today, very little has changed, physically or legally, in this West Virginian water system.</p>
+      </div>
+    </div>
+    <div className="spacer"></div>
+  </div>
+  );
+};
