+++ src/Components/viewer.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default () => <div>
+  viewMaster
+</div>;
